Better handle on errors
# Summary

## Files modified
- isbit_client/client.py -> Adding some logic to give more info on the errors.

## Files added
- tests/test_errors.py -> Some logic on unit tests for the new error logic.

# Description of the problem

##  Intro
The old logic does not have a system for errors that the server can't handle.

## Solution
Adding some status check for the request object, and returning as part of the error the plain text of the request.

## Testing
Asserting that an error will trow when an error code is returned on the header of the request.

# Left TODO
- Test case for error 403
- Assertions on the error message